% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAcuity.R
\name{logmar_to_snellen}
\alias{logmar_to_snellen}
\title{logmar_to_snellen}
\usage{
logmar_to_snellen(logmar, ft = FALSE)
}
\arguments{
\item{logmar}{visual acuity in log MAR (should be a numeric value)}

\item{ft}{if TRUE, return the visual demand (as a string) in snellen feet}
}
\value{
visual acuity in snellen meters (default) or feet
}
\description{
Convert log MAR visual acuity to Snellen Visual acuity
}
\examples{
logmar_to_snellen(1) # 6/60
logmar_to_snellen(1, ft = TRUE) # 20/200
}
\author{
Mark Edosa
}
