% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAcuity.R
\name{min_arc}
\alias{min_arc}
\title{min_arc}
\usage{
min_arc(letter_height, distance)
}
\arguments{
\item{letter_height}{height or size of the letter. This should be in the same unit as the distance}

\item{distance}{distance from the viewer to the letter. This should be in the same unit as the letter.}
}
\value{
Minimum angle of resolution
}
\description{
Determine the minimum angle of resolution
}
\details{
A function that calculates the minimum angle of resolution of a letter at a particular distance
}
\examples{
min_arc(1.5, 110)
min_arc(0.005, 6)
}
\author{
Mark Edosa
}
