% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAcuity.R
\name{snellen_to_logmar}
\alias{snellen_to_logmar}
\title{snellen_to_logmar}
\usage{
snellen_to_logmar(snellen)
}
\arguments{
\item{snellen}{snellen visual acuity in meters or feet}
}
\value{
log MAR visual acuity
}
\description{
Convert Snellen visual acuity to log MAR Visual acuity
}
\details{
This functions log MAR acuity (numeric) from snellen visual acuity (string or numeric)
}
\examples{
snellen_to_logmar("6/60") # 1
snellen_to_logmar("6/9") # 0.1760913
snellen_to_logmar("20/40") # 0.30103
snellen_to_logmar(1) # 0
}
\author{
Mark Edosa
}
