% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualAcuity.R
\name{visual_demand}
\alias{visual_demand}
\title{visual_demand}
\usage{
visual_demand(letter_height, distance, meter = FALSE, feet = FALSE,
  all = FALSE)
}
\arguments{
\item{letter_height}{height or size of the letter. This should be in the same unit as the distance}

\item{distance}{distance from the viewer to the letter, should be in the same unit as the letter.}

\item{meter}{if TRUE, return the visual demand (as a string) in snellen meters}

\item{feet}{if TRUE, return the visual demand (as a string) in snellen feet}
}
\value{
visual acuity in log MAR or snellen or both
}
\description{
Calculates visual demand or threshold for a letter
}
\details{
visual_demand calculates the minimum va required to view a letter of a given size at a given distance
}
\examples{
visual_demand(1.5, 110) # visual demand in log MAR for a letter size of 1.5 inches at a distance of 110 inches
visual_demand(0.005, 6, all = TRUE) # visual demand (in logMAR and snellen) for a 5mm letter viewed at 6 meters
visual_demand(0.005, 6, meter = TRUE) # visual demand (in snellen meters) for a 5mm letter viewed at 6 meters
visual_demand(0.005, 6, feet = TRUE) # visual demand (in snellen feet) for a 5mm letter viewed at 6 meters
visual_demand(0.005, 6, meter= TRUE, feet = TRUE) # visual demand (in both snellen meter and feet) for a 5mm letter viewed at 6 meters
}
\author{
Mark Edosa
}
